rename(product) : 패키지 이동
