+++ src/main/java/kitchenpos/product/infra/JdbcTemplateProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.product.infra;
 
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
