+++ src/test/java/common/ProductFixture.java
@@ -1,7 +1,7 @@
 package common;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductFixture {
